Progress change
I added a health value to the player the value will decrease when the
player is hurt, currently the only added hazard is spikes on the ground.
the level restarts when you fall off the earth that will change later.
There is a button in the middle of stage that currently activates
platforms in order to cross a spike strip.